Refactor and improve existing code
