+++ cli/version.txt
@@ -1,1 +1,1 @@
-0.4.0.-beta.1
+0.4.0-beta.1
